+++ BinanceExchange.API/Models/Response/ExchangeInfoSymbolFilter.cs
@@ -1,9 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Runtime.Serialization;
+﻿using System.Runtime.Serialization;
 using Newtonsoft.Json;
-using BinanceExchange.API.Converter;
-using BinanceExchange.API.Models.Response.Interfaces;
 using BinanceExchange.API.Enums;
 using Newtonsoft.Json.Converters;